+++ src/main/java/de/maxhenkel/easyvillagers/integration/jei/JEIPlugin.java
@@ -4,6 +4,8 @@ import de.maxhenkel.easyvillagers.Main;
 import de.maxhenkel.easyvillagers.blocks.ModBlocks;
 import de.maxhenkel.easyvillagers.integration.jei.breeder.BreederCategory;
 import de.maxhenkel.easyvillagers.integration.jei.converter.ConverterCategory;
+import de.maxhenkel.easyvillagers.integration.jei.incubator.IncubatorCategory;
+import de.maxhenkel.easyvillagers.items.VillagerItem;
 import mezz.jei.api.IModPlugin;
 import mezz.jei.api.registration.IRecipeCatalystRegistration;
 import mezz.jei.api.registration.IRecipeCategoryRegistration;
@@ -17,6 +19,7 @@ import net.minecraft.util.ResourceLocation;
 
 import java.util.ArrayList;
 import java.util.Collection;
+import java.util.Collections;
 import java.util.stream.Collectors;
 
 @mezz.jei.api.JeiPlugin
@@ -24,6 +27,7 @@ public class JEIPlugin implements IModPlugin {
 
     public static final ResourceLocation CATEGORY_BREEDING = new ResourceLocation(Main.MODID, "breeding");
     public static final ResourceLocation CATEGORY_CONVERTING = new ResourceLocation(Main.MODID, "converting");
+    public static final ResourceLocation CATEGORY_INCUBATING = new ResourceLocation(Main.MODID, "incubating");
 
     @Override
     public ResourceLocation getPluginUid() {
@@ -34,12 +38,14 @@ public class JEIPlugin implements IModPlugin {
     public void registerRecipeCatalysts(IRecipeCatalystRegistration registration) {
         registration.addRecipeCatalyst(new ItemStack(ModBlocks.BREEDER), CATEGORY_BREEDING);
         registration.addRecipeCatalyst(new ItemStack(ModBlocks.CONVERTER), CATEGORY_CONVERTING);
+        registration.addRecipeCatalyst(new ItemStack(ModBlocks.INCUBATOR), CATEGORY_INCUBATING);
     }
 
     @Override
     public void registerCategories(IRecipeCategoryRegistration registration) {
         registration.addRecipeCategories(new BreederCategory(registration.getJeiHelpers().getGuiHelper()));
         registration.addRecipeCategories(new ConverterCategory(registration.getJeiHelpers().getGuiHelper()));
+        registration.addRecipeCategories(new IncubatorCategory(registration.getJeiHelpers().getGuiHelper()));
     }
 
     @Override
@@ -55,6 +61,8 @@ public class JEIPlugin implements IModPlugin {
         potions.add(PotionUtils.addPotionToItemStack(new ItemStack(Items.LINGERING_POTION), Potions.WEAKNESS));
         potions.add(PotionUtils.addPotionToItemStack(new ItemStack(Items.LINGERING_POTION), Potions.LONG_WEAKNESS));
         registration.addRecipes(potions, CATEGORY_CONVERTING);
+
+        registration.addRecipes(Collections.singletonList(VillagerItem.getBabyVillager()), CATEGORY_INCUBATING);
     }
 
 }